Convert <Instafeed /> to StaticQuery
