+++ package.json
@@ -9,13 +9,11 @@
   "version": "1.0.0",
   "author": "Matthew Smith",
   "dependencies": {
-    "apollo-client-preset": "^1.0.6",
     "axios": "^0.17.1",
     "chalk": "^2.3.2",
     "disqus-react": "^1.0.5",
     "gatsby": "next",
     "gatsby-image": "next",
-    "gatsby-node-helpers": "^0.3.0",
     "gatsby-plugin-google-analytics": "next",
     "gatsby-plugin-offline": "next",
     "gatsby-plugin-react-helmet": "next",
@@ -30,7 +28,6 @@
     "gatsby-source-filesystem": "next",
     "gatsby-transformer-remark": "next",
     "gatsby-transformer-sharp": "next",
-    "graphql-request": "^1.5.1",
     "jest": "^23.5.0",
     "moment": "^2.21.0",
     "node-sass": "^4.9.3",
