+++ src/components/index.js
@@ -8,11 +8,11 @@ export {default as Footer} from './Footer';
 export {default as Gravatar} from './Gravatar';
 export {default as Icon} from './Icon';
 export {default as Input} from './Input';
+export {default as Instafeed} from './Instafeed';
 export {default as Layout} from './Layout';
 export {default as Link} from './Link';
 export {default as Manifest} from './Manifest';
 export {default as Navbar} from './Navbar';
-export {default as NewInstafeed} from './NewInstafeed';
 export {default as P} from './P';
 export {default as SEO} from './SEO';
 export {default as Tag} from './Tag';
